requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
emoji==0.5.4
ephem==3.7.7.0
future==0.17.1
idna==2.8
pycparser==2.19
PySocks==1.7.0
pyTelegramBotAPI==3.6.6
python-telegram-bot==12.1.0
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.3
clarifai 
